requirements.txt
asyncpg
aiogram
APScheduler
python-decouple
numpy
environ
aiohttp